task 3 updated

+++ task3/not_honest.cpp
@@ -1,4 +1,6 @@
 // решил задачу с доп условием, что не нужно создавать второй массив
+//
+// "более честное" решение можно найти в honest.cpp
 
 #include <iostream>
 
@@ -6,7 +8,7 @@
 
 int main()
 {
-    int array[100000];
+    unsigned long long array[100000];
 
     // инициализируем массив
     #pragma omp parallel for shared(array)
@@ -18,9 +20,9 @@ int main()
     // в задаче известно, что в изначальном массиве хранятся порядковые номера,
     // поэтому для вычисления нового массива данные из старого не нужны
     #pragma omp parallel for shared(array)
-    for (int i = 1; i < 99999; ++i)
+    for (unsigned long long i = 1; i < 99999; ++i)
     {
-        array[i] = (i - 1) * i * (i + 1) / 3.0;
+        array[i] = (i - 1) * i * (i + 1) / 3;
     }
 
     return 0;
